add: ft_memset

+++ ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: jtertuli <jtertuli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/11 09:58:29 by jtertuli          #+#    #+#             */
-/*   Updated: 2025/07/11 11:27:08 by jtertuli         ###   ########.fr       */
+/*   Updated: 2025/07/11 11:34:41 by jtertuli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *b, unsigned int len)
+void	ft_bzero(void *b, size_t len)
 {
 	unsigned char	*ptr;
