+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: jtertuli <jtertuli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/11 08:34:48 by jtertuli          #+#    #+#             */
-/*   Updated: 2025/07/11 11:24:44 by jtertuli         ###   ########.fr       */
+/*   Updated: 2025/07/11 11:35:33 by jtertuli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *dest, int c, unsigned int len)
+#include "libft.h"
+
+void	*ft_memset(void *dest, int c, size_t len)
 {
 	void			*head;
